Add support for group_by query parameter
This commit adds support for the `group_by` query parameter in the `convertUrlToApiCallCodeFence` function. If the `group_by` parameter is present in the URL, it will create a DataFrame using the `response.group_by` data instead of `response.results`. Additionally, if the `group_by` parameter is present, it will display a numeric DataFrame with specific columns selected.

More detailed description:
- Added check for existence of `groupByQuery`
- Updated code to create DataFrame based on presence of `groupByQuery`
- Displayed numeric DataFrame with specific columns when `groupByQuery` is present